Important Scripts Update
Game Manager now contains an int for the number of enemies per level for end game conditions. It Now contains the public function MinusEnemy() which is called in both BaseHealth and TargetAndHealth.
StatManager now contains accessors and mutators for the purpose of adjusting the values in the updated UI. these include the values such as speedCost or speedCount respectfully.
UpgradeValue is located on the Updater child of the TagCollider which is child to WeaponStats on the turrets. This acts as means to assign each weapon type its base stat values. If this can be integrated elsewhere I'm open to changing that.
I removed commented code from the MainTurret script.